On backslash
I found some error on the app, and that's the backslash is shown as a separate codepoint in the orthography, which we actually never use in our orthography. I would like to fix the problem by removing the said symbol. But I don't know whether or not it's something technical. Hopefully not.